nuget pack

+++ vars/virtoModule.groovy
@@ -49,19 +49,26 @@ def call(body) {
 				timestamps {
 					processManifests(false) // prepare artifacts for testing
 
-//					def projs = findFiles(glob:  '**\\*.csproj')
-//					for(int i=0; i<projs.size(); i++){
-//						Packaging.packNuget(this, "\"${projs[i].path}\"")
-//					}
                     String folderPath = "${env.WORKSPACE}\\NuGet"
-                    new File(folderPath).eachFile (FileType.FILES) { file ->
-                        if (file.name.contains('.nupkg')) file.delete()
-                    }
-                    dir(folderPath){
-                        withEnv(["MSBUILD_PATH=${tool 'DefaultMSBuild'}\\msbuild.exe", "PATH+=${env.NUGET}\\nuget.exe"]){
-                            bat "build ${env.WORKSPACE}"
-                        }
-                    }
+					if(new File(folderPath).exists()){
+						new File(folderPath).eachFile (FileType.FILES) { file ->
+							if (file.name.contains('.nupkg')) file.delete()
+						}
+
+						dir(folderPath){
+							def buildFile = new File("build.bat")
+							for (line in buildFile.readLines()) {
+								def res = line =~ /VirtoCommerce\..+\.csproj/
+								if(res.size()>0){
+									bat "${env.NUGET}\\nuget pack \"${env.WORKSPACE}\\${res[0]}\" -IncludeReferencedProjects -Symbols -Properties Configuration=Release"
+								}
+							}
+						}
+					}
+
+					withEnv(["MSBUILD_PATH=${tool 'DefaultMSBuild'}\\msbuild.exe", "PATH+=${env.NUGET}\\nuget.exe"]){
+						bat "build ${env.WORKSPACE}"
+					}
 				}
 			}
